feat: 카드 선택 및 호버 시 CSS 변경
