+++ frontend/src/features/my-calendar/ui/OngoingDiscussion/index.css.ts
@@ -1,4 +1,5 @@
 import { style } from '@vanilla-extract/css';
+import { recipe } from '@vanilla-extract/recipes';
 
 import { vars } from '@/theme/index.css';
 
@@ -9,12 +10,29 @@ export const titleContainerStyle = style({
   backgroundColor: vars.color.Ref.Primary[50],
 });
 
-export const cardStyle = style({
-  padding: vars.spacing[600],
+export const cardStyle = recipe({
+  base: {
+    padding: vars.spacing[600],
 
-  borderRadius: vars.radius[300],
-  border: `1px solid ${vars.color.Ref.Netural[100]}`,
-  backgroundColor: vars.color.Ref.Netural.White,
+    borderRadius: vars.radius[300],
+    border: `1px solid ${vars.color.Ref.Netural[100]}`,
+    backgroundColor: vars.color.Ref.Netural.White,
+
+    cursor: 'pointer',
+
+    ':hover': {
+      backgroundColor: vars.color.Ref.Netural[100],
+      borderColor: vars.color.Ref.Netural[400],
+    },
+  },
+  variants: {
+    selected: {
+      true: {
+        backgroundColor: vars.color.Ref.Netural[100],
+        borderColor: vars.color.Ref.Netural[400],
+      },
+    },
+  },
 });
 
 export const cardTextStyle = style({
